Feat. check session(product api)

+++ fleaewha/src/main/java/kr/or/epro/fleaewha/controller/ProductController.java
@@ -3,7 +3,10 @@ package kr.or.epro.fleaewha.controller;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,6 +14,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.multipart.MultipartFile;
 
 import kr.or.epro.fleaewha.dto.File2;
@@ -21,8 +25,6 @@ import kr.or.epro.fleaewha.service.PostService;
 
 @RestController
 public class ProductController {
-	
-	
 	
 	@Autowired
 	private AWSS3Service service;
@@ -46,23 +48,28 @@ public class ProductController {
 	@PostMapping("/new-product")
 	public String addPost(
 			@RequestPart(value= "file") final List<MultipartFile> multipartFiles,
-			@RequestPart(value= "product") Product2 p
-			) throws Exception {
+			@RequestPart(value= "product") Product2 p,
+			HttpSession session
+	) throws Exception {
+		if(session.getAttribute("id") == null)
+			throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
+		else {
+			postService.addPost(p);
+			
+			String url;
+			for(MultipartFile file : multipartFiles) {
+				File2 file2 = new File2();
+				if(file == multipartFiles.get(0))
+					file2.setType(1);
+				url = "https://fleaewhabucket.s3.ap-northeast-2.amazonaws.com/" + service.uploadFile(file);
+				file2.setProductID(p.getProductID());
+				file2.setFileURL(url);
+				fileService.addFile(file2);
+			}
 		
-		postService.addPost(p);
-		
-		String url;
-		for(MultipartFile file : multipartFiles) {
-			File2 file2 = new File2();
-			if(file == multipartFiles.get(0))
-				file2.setType(1);
-			url = "https://fleaewhabucket.s3.ap-northeast-2.amazonaws.com/" + service.uploadFile(file);
-			file2.setProductID(p.getProductID());
-			file2.setFileURL(url);
-			fileService.addFile(file2);
+			return "post added";
 		}
-	
-		return "post added";
+				
 	}
 
 
@@ -70,36 +77,48 @@ public class ProductController {
     public String updatePost(
     		@RequestPart(value= "file") final List<MultipartFile> multipartFiles,
             @RequestPart(value= "product") Product2 p,
-            @PathVariable int productID
+            @PathVariable int productID,
+            HttpSession session
     ) throws Exception {
     	
-    	p.setProductID(productID);
+    	if(session.getAttribute("id")==null)
+    		throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
+    	else {
+    	  	p.setProductID(productID);
+        	
+        	postService.updatePost(p);
+        
+    		String url;
+    		for(MultipartFile file : multipartFiles) {
+    			File2 file2 = new File2();
+    			if(file == multipartFiles.get(0))
+    				file2.setType(1);
+    			else
+    				file2.setType(0);
+    			url = "https://fleaewhabucket.s3.ap-northeast-2.amazonaws.com/" + service.uploadFile(file);
+    			file2.setProductID(p.getProductID());
+    			file2.setFileURL(url);
+    			fileService.updateFile(file2);
+    		}
+        	
+            return "post updated";
+    	}
     	
-    	postService.updatePost(p);
-    
-		String url;
-		for(MultipartFile file : multipartFiles) {
-			File2 file2 = new File2();
-			if(file == multipartFiles.get(0))
-				file2.setType(1);
-			else
-				file2.setType(0);
-			url = "https://fleaewhabucket.s3.ap-northeast-2.amazonaws.com/" + service.uploadFile(file);
-			file2.setProductID(p.getProductID());
-			file2.setFileURL(url);
-			fileService.updateFile(file2);
-		}
-    	
-        return "post updated";
     }
 
     @DeleteMapping("/products/{productID}")
     public String deletePost(
-            @PathVariable int productID
+            @PathVariable int productID,
+            HttpSession session
     ) throws Exception {
-    	fileService.deleteFile(productID);
-    	postService.deletePost(productID);
-    	return "post deleted";
+    	if(session.getAttribute("id") == null)
+    		throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
+    	else {
+    		fileService.deleteFile(productID);
+        	postService.deletePost(productID);
+        	return "post deleted";
+    	}
+ 
     }
 	
 }
